entrypoint.sh script that starts s6-overlay's `/init` script
If `BITNAMI_APP_DAEMON` is present in the containers environment, the
script  populates any arguments specified on the docker command line
in a `EXTRA_OPTIONS` env variable. The `EXTRA_OPTIONS` variable can
then be used in your service's run script allowing users to specify
additional arguments to your services.

One or more daemon processes can be specified in `BITNAMI_APP_DAEMON`
using the `|` separator. Use this variable in your services run script.
If more than one daemon is specified, then the first one is set as the
main daemon, which can be overridden by specifying an alternate daemon
on the docker command line.

e.g. in `bitnami/wildfly` a user may want to start wildfly in the domain
mode by specifying `domain.sh` on the docker command line instead of the
default standalone mode (`standalone.sh`).

If the user specifies a command which is not listed in `BITNAMI_APP_DAEMON`,
then the containers services are not started and only the user specified
rocess is launched.

e.g. User may want to start the mysql client in the `bitnami/mariadb`
container without starting the mariadb server in the same container.

Lastly, you can override the `entrypoint.sh` script in your derived
image by creating your own.